+++ tests/box_generics.rs
@@ -41,19 +41,25 @@ asteracea::component! {
 }
 
 asteracea::component! {
+	Picky<T: Debug>()()
+
+	box []
+}
+
+asteracea::component! {
 	Custom<T>()()
 
-	box priv a: struct C::<T> <*Custom::<T>>
+	box priv a: struct C::<T> <*Boxed::<T>>
 }
 
 asteracea::component! {
 	CustomClaused<T: Debug>()()
 
-	box priv a: struct CC::<T: Debug> <*CustomClaused::<T>>
+	box priv a: struct CC::<T: Debug> <*Picky::<T>>
 }
 
 asteracea::component! {
 	CustomWhered<T> where T: Debug, ()()
 
-	box priv a: struct CW::<T> where T: Debug; <*CustomWhered::<T>>
+	box priv a: struct CW::<T> where T: Debug; <*Picky::<T>>
 }